管理者編集をAdministratorControllerに分離
