+++ laravel-project/app/Http/Controllers/Admin/UserController.php
@@ -33,73 +33,6 @@ class UserController extends Controller
         return view('admin.index', compact(['users', 'name', 'email']));
     }
 
-    public function editAdmin()
-    {
-        if (Gate::denies('admin.authority')) {
-            abort(403);
-        }
-
-        $admin = Auth::user();
-
-        return view('admin.edit', compact(['admin']));
-    }
-
-    public function confirmAdmin(Request $request)
-    {
-        if (Gate::denies('admin.authority')) {
-            abort(403);
-        }
-
-        $adminId = Auth::id();
-
-        $formData = $request->validate([
-            'name' => 'required|string|max:255',
-            'email' => ['required', 'email', Rule::unique('users')->ignore($adminId), 'max:255'],
-            'password' => 'nullable|min:8|confirmed',
-        ]);
-
-        return view('admin.confirm', compact(['formData']));
-    }
-
-    public function updateAdmin(Request $request)
-    {
-        if (Gate::denies('admin.authority')) {
-            abort(403);
-        }
-
-        $adminId = Auth::id();
-
-        $input = $request->only(['name', 'email', 'password']);
-
-        $validator = Validator::make($input, [
-            'name' => 'required|string|max:255',
-            'email' => ['required', 'email', Rule::unique('users')->ignore($adminId), 'max:255'],
-            'password' => 'nullable|min:8',
-        ]);
-        if ($validator->fails()) {
-            return redirect()->route('admin.edit')
-                ->withInput()
-                ->withErrors($validator);
-        }
-
-        if ($request->input('back') == 'back') {
-            return redirect()->route('admin.edit')
-                ->withInput();
-        }
-
-        $formData = $validator->validated();
-
-        if (is_null($formData['password'])) {
-            unset($formData['password']);
-        }
-
-        $user = User::find($adminId);
-        $user->fill($formData);
-        $user->save();
-
-        return redirect()->route('admin.index');
-    }
-
     public function create()
     {
         if (Gate::denies('admin.authority')) {
